pyproject.toml
[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "pyspectral"
version = "0.1.0"
description = "Machine learning for processing spectral imaging"
license = { file = "LICENSE" }
readme = "README.md"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License"
]
dependencies = [
    "beartype>=0.22.5",
    "ipympl>=0.9.8",
    "jaxtyping>=0.3.3",
    "loguru",
    "marimo>=0.19.10",
    "matplotlib>=3.10.6",
    "mkdocs",
    "nbconvert>=7.17.0",
    "nbformat>=5.10.4",
    "numpy>=2.3.2",
    "playwright>=1.58.0",
    "polars>=1.33.0",
    "python-dotenv",
    "scikit-image>=0.25.2",
    "scikit-learn>=1.7.1",
    "scipy-stubs>=1.16.3.1",
    "torch>=2.8.0",
    "torchsummary>=1.5.1",
    "torchvision>=0.23.0",
    "tqdm",
    "ty>=0.0.16",
]
requires-python = "~=3.13.0"

[dependency-groups]
dev = [
    "pytest",
    "ruff",
    "ipykernel>=6.30.1",
    "mypy>=1.18.2",
]

[tool.uv.sources]
torch = [
    { index = "pytorch-cu128", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]
torchvision = [
    { index = "pytorch-cu128", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]

[[tool.uv.index]]
name = "pytorch-cu128"
url = "https://download.pytorch.org/whl/cu128"
explicit = true

[project.scripts]
pyspec = "pyspectral.cli:main"

# ============================================================
# RUFF CONFIGURATION 
# ============================================================

[tool.ruff]
line-length = 88
src = ["pyspectral"]
include = ["pyproject.toml", "pyspectral/**/*.py"]

[tool.ruff.lint]
extend-select = ["I"]  # import sorting
ignore = ["F722"]  # for fixing JAXtyping

[tool.ruff.lint.isort]
known-first-party = ["pyspectral"]
combine-as-imports = true
force-sort-within-sections = true

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

# ============================================================
# type linting settings
# ============================================================
[tool.mypy]
warn_unused_configs = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_return_any = true
no_implicit_optional = true
check_untyped_defs = false
disallow_untyped_defs = true
strict_equality = true

[[tool.mypy.overrides]]
module = ["matplotlib.*", "torch.*", "sklearn.*"]
ignore_missing_imports = true

[tool.zuban]
strict = true
disallow_untyped_defs = true
warn_unreachable = true

[tool.basedpyright]
include = ["pyspectal"]
stubPath = "stubs/"
defineConstant = { DEBUG = true }
pythonVersion = "3.13"
pythonPlatform = "Windows"
reportAny = false
autoImportCompletions = true
useLibraryCodeForTypes = true
reportMissingTypeArgument = false
reportMissingTypeStubs = false
reportUnknownArgumentType = false
reportExplicitAny = false
reportUnknownMemberType = false
reportIgnoreCommentWithoutRule = false
reportUnknownParameterType = false
reportUntypedFunctionDecorator = false
reportConstantRedefinition = true
reportUnknownVariableType = false
reportUnusedCallResult = false
reportUnusedImport = "none"
reportMissingImports = "error"
reportUnusedVariable = "hint" 
# Null safety
reportOptionalSubscript = true
reportOptionalMemberAccess = true
reportOptionalCall = true
reportOptionalIterable = true
reportOptionalContextManager = true
reportOptionalOperand = true

# ============================================================
# EXCLUSIONS
# ============================================================
exclude = [
    "**/__pycache__",
    "**/node_modules",
    ".git",
    ".mypy_cache",
    ".pyright_cache",
    ".ruff_cache",
    ".pytest_cache",
    ".venv",
    "venv",
    "env",
    "logs",
    "output",
    "data",
    "build",
    "dist",
    "*.egg-info",
]

venvPath = "."
venv = ".venv"

# ============================================================
# PYTEST CONFIGURATION
# ============================================================
[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
]

# ============================================================
# COVERAGE CONFIGURATION
# ============================================================
[tool.coverage.run]
source = ["."]
branch = true  # STRICT: Enable branch coverage
omit = [
    "*/tests/*",
    "*/test_*.py",
    "*/__pycache__/*",
    "*/.venv/*",
    "*/venv/*",
    "*/scripts/*",
]

[tool.coverage.report]
precision = 2
show_missing = true
skip_covered = false
fail_under = 30  # Require 30% coverage
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@abstractmethod",
    "@overload",
]
